requirements.txt
bcrypt
blinker
cachelib
click
colorama
dnspython
Flask
Flask-Caching
Flask-Cors
gunicorn
itsdangerous
Jinja2
MarkupSafe
numpy
opencv-python
Pillow
PyJWT
pymongo
python-dotenv
six
Werkzeug